Remove duplicate styles
