+++ front/app/containers/HomePage/SignedInHeader/CompleteProfileStep.tsx
@@ -1,8 +1,10 @@
 import React from 'react';
 import { isNilOrError } from 'utils/helperUtils';
-
+import OnboardingStep from './OnboardingStep';
+import SkipButton from './SkipButton';
+import AcceptButton from './AcceptButton';
+import { Left, Right, Text, HeaderContent, Icons, StyledAvatar } from './';
 // components
-import Avatar from 'components/Avatar';
 import { Icon } from '@citizenlab/cl2-component-library';
 
 // services
@@ -15,117 +17,9 @@ import { useIntl } from 'utils/cl-intl';
 
 // style
 import styled from 'styled-components';
-import { media, fontSizes, isRtl } from 'utils/styleUtils';
 
 // hooks
 import useAuthUser from 'hooks/useAuthUser';
-import OnboardingStep from './OnboardingStep';
-import SkipButton from './SkipButton';
-import AcceptButton from './AcceptButton';
-
-const contentTimeout = 350;
-const contentEasing = 'cubic-bezier(0.19, 1, 0.22, 1)';
-const contentDelay = 550;
-
-const HeaderContent = styled.div`
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding-top: 20px;
-  padding-bottom: 20px;
-  padding-left: 75px;
-  padding-right: 75px;
-  overflow: hidden;
-
-  ${isRtl`
-    flex-direction: row-reverse;
-  `}
-
-  &.content-enter {
-    opacity: 0;
-
-    &.content-enter-active {
-      opacity: 1;
-      transition: all ${contentTimeout}ms ${contentEasing} ${contentDelay}ms;
-    }
-  }
-
-  &.content-enter-done {
-    opacity: 1;
-  }
-
-  &.content-exit {
-    opacity: 1;
-    transition: all ${contentTimeout}ms ${contentEasing};
-
-    &.content-exit-active {
-      opacity: 0;
-    }
-  }
-
-  &.content-exit-done {
-    display: none;
-  }
-
-  h2 {
-    color: #fff;
-    font-size: ${fontSizes.xxl}px;
-    line-height: 33px;
-    font-weight: 400;
-  }
-
-  ${media.tablet`
-    padding-left: 30px;
-    padding-right: 30px;
-  `}
-
-  ${media.tablet`
-    flex-direction: column;
-    align-items: stretch;
-    justify-content: center;
-    padding-left: 15px;
-  `}
-`;
-
-export const Left = styled.div`
-  display: flex;
-  align-items: center;
-  margin-right: 60px;
-
-  ${isRtl`
-    flex-direction: row-reverse;
-  `}
-
-  ${media.tablet`
-    margin-right: 0;
-    flex-direction: column;
-    align-items: flex-start;
-  `}
-
-  ${media.phone`
-    align-items: center;
-    margin-bottom: 30px;
-  `}
-`;
-
-export const Icons = styled.div`
-  display: flex;
-  margin-right: 30px;
-
-  ${isRtl`
-    margin-right: 0px;
-    margin-left: 30px;
-  `}
-
-  ${media.phone`
-    margin-right: 0;
-  `}
-`;
 
 const CompleteProfileIcon = styled(Icon)`
   width: 48px;
@@ -133,39 +27,6 @@ const CompleteProfileIcon = styled(Icon)`
   margin-left: -3px;
   margin-top: 3px;
   opacity: 0.5;
-`;
-
-export const Text = styled.div`
-  ${isRtl`
-    direction: rtl;
-  `}
-
-  ${media.phone`
-    text-align: center;
-  `}
-`;
-
-export const Right = styled.div`
-  display: flex;
-
-  ${isRtl`
-    flex-direction: row-reverse;
-  `}
-
-  ${media.phone`
-    flex-direction: column;
-  `}
-`;
-
-export const AvatarAndShield = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
-
-export const StyledAvatar = styled(Avatar)`
-  margin-right: -3px;
-  z-index: 2;
 `;
 
 interface Props {
